Altered Outcome data to handle Baseline and Q2,3,4. Handles Grades for multiple electives
Added code to change Baseline to another report period when that report
period is not present. For attendance outcomes, the code will use the
date associated with the outcome to find the report period during which
the Baseline outcome was entered. If there is not an outcome entered
for that report period, then the baseline outcome is changed to that
report period. For grades outcomes, there is not a date associated with
the outcome in the data (have no idea why not) so Baseline grades
outcomes are replaced by 1st Grading Period if there is not an entry
for the first grading period.

**Note** There are some outcome entries that have nonsense report
periods entered, such as “First Semester”. We might want to make an
excel export for errors such as this and duplicated outcomes to report
to GC’s.

Added code to create variables for each of the core academic subjects
for each quarter (ex. “Q_1 Science” ). Avgrade1 is the average grade
for these subjects in quarter 1, Avgrade2 for quarter 2, and so on.
Science, Math, Lang. Arts, etc. are the average grades in Science,
Math, Lang. Arts over all existing quarters.

Added commented out code that replaces the subject of a grade outcome
with the text in the Note if the note exists. It’s commented out
because we might want to edit it and have a different method, but it’s
a first idea.